[Add] SmallFavorButton 구현
